requirements.txt
sentence-transformers 
faiss-cpu 
openai 
transformers
psycopg2
sqlglot
SQLAlchemy